Software Security Project
Design, develop, test and document a secure solution to the problem
Host the solution in a Google colab and in a Lambda function, accessed through an API endpoint.
1. Design: How solution to work.
2. Develop: Create solution, and a testing .
3. Document: Instructions on how to use solution.
4. Secure the Lambda function: Code is secure, meaning it's protected from unauthorized
access and potential threats.
5. Create, secure and test the API Endpoint: A way for solution to be accessed through an API
endpoint.
THE PROBLEM
When people need to relocate to a new area, they may need to find schools within a reasonable
distance of where they would live. This might determine their choice of location. If the process
was automated, it would be quick to use and the information could come from one place, rather
than a number of places.
Potentially, if a list could be stored, then it could be edited to reduce the list to a selection of
schools that might be used in decision making about where to locate.
USER STORIES
AS a user
WHEN I want to find the nearest schools to a given location
THEN I will be able to run a function in a Google Colab that will ask Google maps to send a list of
nearby schools
AND I will be shown a list of schools in that area
AND this information will also be sent to a serverless function on AWS and saved in a schools file
on S3
AS a user
WHEN I want to see all the schools in the file on AWS S3
THEN I will be able to run a function in a Google Colab that will send a request to get the full list,
AND I will see the list printed in my output